Add files via upload
Simple List, Slice Strings, Python

Is is incredibly useful wwhen you want to work with text. For instance, logs. You can use this to parse logs by slicing the logs at specific points on the log currently loaded to a variable. Note this is only useful if the strings are of the same size, IE IP Addresses, MAC Addresses, phone numbers, zipcodes, etc