Add files via upload
Zip Function, FOR Loop with variable with multiple Lists, Python

Zip function is used when the index (relative position)
of the data in question DOES NOT matter. Please note that
zip with different size lists will stop after the shortest list 
runs out of items.

E.G If the zip function is used, and list 1 has twenty entries,
and list 2 has twelve entries, zip function will stop for loop
after the 12th entry is looped.

This is incredibly useful when working with an unknown number of items or variables, such as log data where items may have some fields populated, but not others.